simplify utils app objects
