+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>edu.baylor.ecs.cloudhubs</groupId>
     <artifactId>prophet-dto</artifactId>
-    <version>0.0.6</version>
+    <version>0.0.7</version>
 
     <properties>
         <maven.compiler.source>1.8</maven.compiler.source>
